requirements.txt
Django==1.7.3
elasticsearch==1.4.0
djangorestframework==3.0.5
mysql==0.0.1
MySQL-python==1.2.5